requirements.txt
aiobotocore==2.5.4
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
botocore==1.31.17
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chocolate==0.0.2
colorama==0.4.6
cryptography==41.0.3
csvwlib==0.3.2
decorator==5.1.1
duckdb==0.8.1
elementpath==4.1.5
executing==1.2.0
falcon==3.1.1
frozenlist==1.4.0
fsspec==2023.6.0
gcsfs==2023.6.0
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
html5lib==1.1
icecream==2.1.3
idna==3.4
ipython==8.14.0
isodate==0.6.1
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
JPype1==1.4.0
jsonpath-python==1.0.6
jsonpickle==3.0.2
kglab==0.6.6
language-tags==1.2.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
morph_kgc==2.6.2
multidict==6.0.4
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
openai==0.27.9
owlrl==6.0.2
oxrdflib==0.3.6
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
prettytable==2.5.0
prompt-toolkit==3.0.39
protobuf==4.24.2
pslpython==2.4.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pynvml==11.5.0
pyoxigraph==0.3.19
pyparsing==3.1.1
pyshacl==0.23.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyvis==0.3.2
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3fs==2023.6.0
scikit-learn==1.1.3
scipy==1.11.2
six==1.16.0
sql-metadata==2.8.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.14.0
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
urlpath==1.2.0
wcwidth==0.2.6
webencodings==0.5.1
wrapt==1.15.0
yarl==1.9.2
